Separate `usePsbtValidate` hook from view component
